Front-end logic
In short, I made the logic on the client side. In detail, quite a lot was done, the main thing is a redux basket that is synchronized with a cookie (redux is also available via getServerSideProps, etc.). I also made the logic and validation for the checkout form, now there is validation (unfortunately, it is not the most detailed, but it is in the process). I also decided to add a little beauty and added a custom cursor via redux-animated-cursor. In fact, a lot has been done, but there is no point in writing it here(its need to see).