Changed API route
Made some changes in the API route. The main thing is an email template for the client and owner of the site. Now if some people buy some products and type in the checkout form real email it receives to email version of the order with some info about an order.